dataset_loader :Add plots with seaborn for leaning matter. Add of a dictionnary to get the meaning of each feature if necessary.
modelisation_boston_house_princing:
Add of a preprocessing class to build the training and testing dataset.
Add some verification on the dataset (Nan and duplicated rows.)
In the model BOSTON_HOUSE_PRINCING class, add two type of regressor,
the Random Forest Regressor and the GradientBoosting Regressor.
For learning matter, a function to create the plots with seaborn has been added.
Functions to create and plots learning curve and validation curve have been added.
Comparaison between the 3 regressor is done in the main function.